Multiple linear implemented, need to fix issues after calling evaluate.
Final trace of the error is in line 19. Although it is entirely possible that the error is due earlier processes in the code.

Main function implemented - imports sklean toy dataset, applies cross validation, trains the model on the training set and evaluates on the testing set.

CV is not necessary for the assignment, but I want some measure to see how good the model is (if it is any good).